Added check for winnings
Added a check to see if one of the players has won. All sides and diagonals around the player's chosen cell are checked for compliance with the player's sign. If the previous (or next) cell was also chosen by the player, then it is checked in a reversible loop until the number of cells passed is equal to the maximum. If the number of squares passed is the maximum, then this player wins.